Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/https-header-subresource.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/testharness.js"></script>
    <script src="/js-test-resources/testharnessreport.js"></script>
    <script>
    if (window.testRunner)
      testRunner.setCanOpenWindows();

    var t = async_test("Verify that subresources do not receive an 'Upgrade-Insecure-Requests' header.");
    fetch("resources/echo-https-header.pl").then(t.step_func(function(res) {
        return res.json();
    })).then(t.step_func(function(json) {
        assert_equals(json.header, "");
        t.done();
    }));
    </script>
</head>
<body>
</body>
</html>
